refactor(CatListPage): break down the page to simplify the code read
