+++ src/pages/CatListPage/index.tsx
@@ -1,0 +1,2 @@
+import CatListPage from './CatListPage'
+export default CatListPage
